Move bottom structs to better places
