+++ aggregator/rpc_server.go
@@ -1,0 +1,49 @@
+package aggregator
+
+import (
+	"context"
+	"net/http"
+	"net/rpc"
+
+	"github.com/Layr-Labs/eigensdk-go/crypto/bls"
+	blsagg "github.com/Layr-Labs/eigensdk-go/services/bls_aggregation"
+	sdktypes "github.com/Layr-Labs/eigensdk-go/types"
+)
+
+func (agg *Aggregator) startServer(ctx context.Context) {
+	err := rpc.Register(agg)
+	if err != nil {
+		agg.logger.Fatal("Format of service TaskManager isn't correct. ", "err", err)
+	}
+	rpc.HandleHTTP()
+	err = http.ListenAndServe(agg.serverIpPortAddr, nil)
+	if err != nil {
+		agg.logger.Fatal("ListenAndServe", "err", err)
+	}
+}
+
+type SignedTaskResponse struct {
+	TaskResponse TaskResponse
+	BlsSignature bls.Signature
+	OperatorId   sdktypes.OperatorId
+}
+
+
+// rpc endpoint which is called by operator
+// reply doesn't need to be checked. If there are no errors, the task response is accepted
+// rpc framework forces a reply type to exist, so we put bool as a placeholder
+func (agg *Aggregator) ProcessSignedTaskResponse(signedTaskResponse *SignedTaskResponse, reply *bool) error {
+	agg.logger.Infof("Received signed task response: %#v", signedTaskResponse)
+	taskIndex := signedTaskResponse.TaskResponse.TaskIndex()
+
+	taskSignature := blsagg.NewTaskSignature(
+		taskIndex,
+		signedTaskResponse.TaskResponse,
+		&signedTaskResponse.BlsSignature,
+		signedTaskResponse.OperatorId,
+	)
+
+	err := agg.blsAggregationService.ProcessNewSignature(context.Background(), taskSignature)
+
+	return err
+}
